# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Readify API Documentation",
        "description": "API documentation for Readify application",
        "contact": {
            "email": "muhammadzelvan263@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Readify API Server"
        }
    ],
    "paths": {
...
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get list of reservations",
                "operationId": "b8d33d9a2963a7cd60615008cd626a56",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "book_id",
                        "in": "query",
                        "description": "Filter by book ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reservations retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "operationId": "36133bb1e72eeb990aa0187b2afddb2b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "book_id",
                        "in": "query",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reservation_date",
                        "in": "query",
                        "description": "Reservation date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "expiry_date",
                        "in": "query",
                        "description": "Expiry date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Reservation created successfully"
                    }
                }
            }
        },
        "/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get a specific reservation",
                "operationId": "780bd81b9af93bc3af711c4827b90955",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation retrieved successfully"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update a reservation's status",
                "operationId": "0058252572cd4cfc9136c717907ed312",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the reservation",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "canceled",
                                "completed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation status updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation",
                "operationId": "eb324f4396575f61a9d1d3a67733847e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully"
                    }
                }
            }
        }
    },
...
}